Update to Travis staged build.
These changes are because the exisitng build has these drawbacks:
Code in the after_success: part of the build does not fail the
build.  This makes it difficult to verify that (i) the doc build
works and (ii) that doctest etc. still passes.
Doc building runs in the global environment, and is thus affected
by the surroundings.